[dataset] move out mongo helpers in a specific module
This will helper organize possible helpers for other possible future
backends.
